Day05/Intelligent Travel Agent/requirements.txt
langchain
requests
openai
duckduckgo-search
python-dotenv
streamlit 
transformers
huggingface_hub
torch